Fixups and stubbing

+++ src/JSystem/J2DGraph/J2DAnmLoader.cpp
@@ -107,7 +107,7 @@ void* J2DAnmKeyLoader_v15::load(void const* p_data) {
         default:
             break;
         }
-        dataPtr = (J3DAnmDataBlockHeader*)((s32)dataPtr + dataPtr->mNextOffset);
+        dataPtr = (J3DAnmDataBlockHeader*)((uintptr_t)dataPtr + dataPtr->mNextOffset);
     }
 
     return mpResource;
@@ -148,7 +148,7 @@ void J2DAnmKeyLoader_v15::setResource(J2DAnmBase* p_anm, void const* p_data) {
         default:
             break;
         }
-        dataPtr = (J3DAnmDataBlockHeader*)((s32)dataPtr + dataPtr->mNextOffset);
+        dataPtr = (J3DAnmDataBlockHeader*)((uintptr_t)dataPtr + dataPtr->mNextOffset);
     }
 }
 
@@ -295,7 +295,7 @@ void* J2DAnmFullLoader_v15::load(void const* p_data) {
         default:
             break;
         }
-        dataPtr = (J3DAnmDataBlockHeader*)((s32)dataPtr + dataPtr->mNextOffset);
+        dataPtr = (J3DAnmDataBlockHeader*)((uintptr_t)dataPtr + dataPtr->mNextOffset);
     }
     return mpResource;
 }
@@ -334,7 +334,7 @@ void J2DAnmFullLoader_v15::setResource(J2DAnmBase* p_anm, void const* p_data) {
         default:
             break;
         }
-        dataPtr = (J3DAnmDataBlockHeader*)((s32)dataPtr + dataPtr->mNextOffset);
+        dataPtr = (J3DAnmDataBlockHeader*)((uintptr_t)dataPtr + dataPtr->mNextOffset);
     }
 }
